requirements.txt
python-telegram-bot
google-generativeai
requests
replicate
langchain-community